requirements.txt
crewai
crewai_tools
langchain==0.1.20
langchain-community==0.0.38
langchain-google-genai==1.0.4
load_dotenv
youtube_search